Move into another OIDC test
