+++ ci/tests/puppeteer/scenarios/oidc-login-jwt-hmac-metadata-cors/init.sh
