+++ ci/tests/puppeteer/scenarios/oidc-login-jwt-hmac-metadata-cors/services/Sample-1.json
